+++ src/components/layout/StakingDashboard/card/BaseCard.jsx
